refactor: 골 기간에 대한 필드 Value Object로 수정

+++ src/main/java/com/backend/blooming/goal/presentation/dto/response/GoalResponse.java
@@ -15,7 +15,7 @@ public record GoalResponse(
 
         @DateTimeFormat(pattern = "yyyy-MM-dd")
         String endDate,
-        int days,
+        long days,
         Long managerId,
         List<Long> teamUserIds
 ) {
